preprocessor/lexer/parser/: Improved support for external syntax highlighting (#89)
Improv: last changes for syntax highlighter done in view_patternEditor. Other fixes/changes include:
Fix: lexer was reading include folders past the end of file when include was the last token in file.
Highlighter: parser.cpp now provides better information about the identifier types.
Highlighter: Some information coudlnt be obtained because the parser created new preprocessors and new token streams to parse imported files which were later destroyed when no longer in use. In particular the namepaces encountered during parsing were only stored in the asts as parts of the names of functions and custom types. I resolve this by letting the preprocessor store the namespaces and saving them to a vector. Because new preprocessors were being created (and destroyed) I also open the original preprocessor used before the parser to copy the information from the new preprocessors so that it can be accessed when highlighting.
Highlighter: I took some basic timing measurements and the slowest part by far far was the lexing-preprocessing-parsing, that is, the creation of the token stream and the ASTNode vector. Currently, when any change in the text editor occurs a background task is created to parse the code. This is needed to deal with in and out variables, but I thought that instead of  redoing it all over again I could just use the results obtained in the task. Unfortunately the ast is not savedand it is thrown away, so I change the code to make sure that m_currAST gets loaded with that result.
  Highlighter: template variables were missing in the list of identifier types so it was added as well.