preprocessor: Undo import-like syntax support for includes. (#139)
Just like imports can use include-like syntax I tried to make includes support import-like syntax. That way the choice would depend less on the format and more on the capabilities of each implementation. I got the distinct impression that it wasn't a good idea and on top of that I screwed up and made includes unusable. Luckily Werwolv came to the rescue and fixed the bad code but what's left is hard to read because it still contains code to support import syntax while not supporting it anymore.

So I've decided to write code that support only the original syntax of includes so these are correct:
```cpp
#include "std/io.pat"
#include "std/io"
#include "std/io"
#include "std\\io" //on windows only
#include <std/io.pat>
```
and so that these forms don't work
```cpp
#include "std.io"
#include <std.io>
```
and so on.
Hopefully this makes the code make a bit more sense and work as expected. At least it has in all the tests I've done.