preprocessor: Fixed pragma once with imports and includes at the same time (#137)
After merging the code that makes includes and imports affect each other under #pragma once directive the import unit test on the Pattern Language repo started to fail. The reason was that imports with empty aliases were being treated as includes and processed before the imports were parsed, in effect, blocking themselves. To fix it, split imports and includes into their own separate lists and only pass the parser the list of includes.