evaluator: Fixed breakpoints becoming disabled after first hit. (#131)
* fix: breakpoints become disabled after first hit.

If you set a breakpoint the first time you run the pattern it gets hit. After that subsequent runs will fail to hit the breakpoint.

This bug was caused by two factors.
- When breakpoint hits evaluation stops before the line with the breakpoint is evaluated. In order to avoid hitting the breakpoint again when evaluation is continues the breakpoint needs to be disabled. Once the line is evaluated the breakpoint needs to be enabled again, but the previous code didn't do that.
- Detecting if the line was being evaluated used insufficient information to perform said detection. Because of the way AST nodes are assigned coordinates, Nodes that lie in other lines iof source code may and up sharing the same line number as the Nodes in the breakpoint. To tell them apart one must also use columns

Both problems were resolved. When evaluation of another line occurs the breakpoint is re-enabled. Note that at this point the full evaluation of the line may not have finish yet. This is of no concern as the second problem prevents triggering the breakpoint erroneously. By using the column of the position of the node we ensure that only valid evaluations of the node trigger the breakpoints.

* fix: Corrected terrible implementation of good fix. The fixed required the comparison of the column location of the node to see if it was part of the actual code on the line. The problem was that it was computing the lines of the string content in every node comparison instead of just once before evaluation starts. For large input files that meant a considerable slow-down which can be easily avoided.

* fix: remaining issues with breakpoints. Forgot to reset the line lengths and the lastPausedLine variables on every new evaluation.lastPauseLine is necessary if e.g. You press f5 while paused at a breakpoint so the evaluation hits the breakpoint again.

More importantly the new code now handles one-step debug a lot better than before. The way nodes get coordinates associated with them makes it impossible touse line number information to determine when the next pause will occur. That was fixed for breakpoints, and the new code also fixes it for one-step debugging as well. Additionally, only source code will be subject to pause. Previously, if you placed a breakpoint at a location that happened to be evaluated in an imported file, it was hitting the imported file too. Also stepping through code would go inside imported files which is not what we want most likely given the fact that only source code is displayed while debugging.

* Changes suggested by the reviewer.