fix: imported patterns used for inheritance were not highlighted. (#185)
The problem was that imported files didn't have token sequences to obtain the UDT variables. The fix was to create maps from the file name to the token sequence and then process each imported file to obtain all the variables needed. Function variables are skipped since they can be part of the code.
There are also some minor code style corrections and a fix in the text editor where the last line of a selection was not being deleted.