evaluator: Fix type name for non-placement template variables (#208)
evaluator: fix type name for non-placement template variablesFix type name for non-placement template variables

Fix the result of the type name for non-placement variables with template types.
We still need to evaluate the arguments of the template before setting the type name.
For typeNameOf the type, we can skip creating the pattern as only the arguments matter.

Co-authored-by: Nik <werwolv98@gmail.com>