Change to make the upcoming pt in imhex repo compile without errors. (#217)
Change to make the upcoming pt in imhex repo compile without errors. The new changes made there cause a mussing constructor error on the errors class because it is missing a default constructor or at least the error goes away when the default constructor is added. If there is a better way to handle this let me know.